% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_module.R
\name{DESeq2_module}
\alias{DESeq2_module}
\title{DESeq2 module taking metadata_csv nad raw count data frame to run DE analysis}
\usage{
DESeq2_module(
  count_data,
  anno_tb = NULL,
  tpm_tb = NULL,
  tag = NULL,
  metadata_csv = NULL,
  metadata_design = NULL,
  control_reference = NULL,
  output_dir = NULL,
  delim_samples = "\\\\."
)
}
\arguments{
\item{count_data}{object in matrix format (integers) with rownames as genes and colnames as sample IDs}

\item{anno_tb}{tibble of ensembl_gene_id - external_gene_id mappings for annotation}

\item{tpm_tb}{tibble of tpms to append to results for output}

\item{tag}{string used to prefix output}

\item{metadata_csv}{path to file holding metadata}

\item{metadata_design}{string of design using colnames of metadata_csv}

\item{control_reference}{string indicating the 'control' intgroup for last element of design}

\item{output_dir}{path to where output goes}

\item{delim_samples}{character to delimit sample IDs (default: ".")}

\item{ens_version}{ENSEMBL version to use}

\item{org_prefix}{organism prefix from ENSEMBL}
}
\value{
nz_fc_co raw count object with no lines summing to zeros
}
\description{
DESeq2 module taking metadata_csv nad raw count data frame to run DE analysis
}
