% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupradar.R
\name{dupradar_run}
\alias{dupradar_run}
\title{Function to run fGSEA using the mSIGDB database from Broad}
\usage{
dupradar_run(gtf, paired, stranded = "0", threads)
}
\arguments{
\item{gtf}{path/to/GTF.gtf}

\item{paired}{boolean of whether BAM data is paired}

\item{stranded}{string '0', '1', '2', indicating 'unstranded', 'forward', reverse' respectively}

\item{threads}{total threads used to run}
}
\value{
RData object of dupRadar outputs in list
}
\description{
Function to run fGSEA using the mSIGDB database from Broad
}
