% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea.R
\name{fgsea_plot}
\alias{fgsea_plot}
\title{Function to run fGSEA using the mSIGDB database from Broad}
\usage{
fgsea_plot(fgsea_res_tb, msigdb_cat)

fgsea_plot(fgsea_res_tb, msigdb_cat)
}
\arguments{
\item{fgsea_res_tb}{tibble of results from fgsea::fgsea}

\item{msigdb_cat}{one of 'c("H", paste0("C", c(1:7)))',
see: gsea-msigdb.org/gsea/msigdb/collections.jsp}

\item{res}{output table from 'DESeq2_module()' (all genes not just sig)}

\item{sig_res}{output table from 'DESeq2/limma/edger_module()' or combination thereof (used for table output)}

\item{msigdb_species}{one of msigdbr::msigdbr_show_species(), default:"Homo sapiens"}

\item{gene_col}{the name of the column in tibble with gene names found in pathways, set to "rownames" if they are the rownames (default)}

\item{rank_col}{the name of the column in tibble which ranks genes for fgsea (default: "stat" for DESeq2 results; limma - use "t", edgeR - unsure)}

\item{padj}{the significance threshold}

\item{output_dir}{path to where output goes}

\item{tag}{string used to prefix output}

\item{contrast}{string to define the contrast being made, tags output}
}
\value{
msigdb_fgsea object

msigdb_pathlist list object
}
\description{
Function to run fGSEA using the mSIGDB database from Broad

fgsea plotting
}
