% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea.R
\name{fgsea_plotting}
\alias{fgsea_plotting}
\title{fgsea plotting}
\usage{
fgsea_plotting(fgsea_res_tb, msigdb_cat)
}
\arguments{
\item{fgsea_res_tb}{tibble of results from fgsea::fgsea}

\item{msigdb_cat}{one of 'c("H", paste0("C", c(1:7)))',
see: gsea-msigdb.org/gsea/msigdb/collections.jsp}
}
\value{
msigdb_pathlist list object
}
\description{
fgsea plotting
}
