% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne-ap_viper.R
\name{fgsea_ssgsea_msviper}
\alias{fgsea_ssgsea_msviper}
\title{Run fgsea on each regulon set individually}
\usage{
fgsea_ssgsea_msviper(
  mra,
  mra_stat = "p.value",
  genome_prefix,
  msigdb_species = "Homo sapiens",
  msigdb_cat = "H",
  tag,
  exprdh,
  tx2gene,
  gene_col = "external_gene_name",
  output_dir,
  sig_val = 0.1,
  per_regulon = NULL,
  phenos
)
}
\arguments{
\item{mra}{msviper class object with es, ledge}

\item{mra_stat}{statistic used to rank data, comes from msViper}

\item{genome_prefix}{string of a genome in biomart$dataset, suffixed with '_gene_ensembl'}

\item{msigdb_species}{one of msigdbr::msigdbr_species(), default:"Homo sapiens"}

\item{msigdb_cat}{one of 'c("H", paste0("C", c(1:7)))', see: gsea-msigdb.org/gsea/msigdb/collections.jsp}

\item{tag}{string used to prefix output}

\item{exprdh}{expression data}

\item{tx2gene}{annotation data}

\item{gene_col}{the name of the column in tibble with gene names found in pathways, set to "rownames" if they are the rownames (default)}

\item{output_dir}{path to where output goes}

\item{sig_val}{significance used in analysis}

\item{per_regulon}{flag to indicate if fgsea is at collective regulon level,
or if each regulon and the genes it is co-expressed with are used}

\item{phenos}{metadata tibble}
}
\value{
sure
}
\description{
Run fgsea on each regulon set individually
}
