% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_RNAseq_output.R
\name{get_metadata}
\alias{get_metadata}
\title{Make metadata from CSV input, including generation of 'path' for abundance.h5 kallisto ouput
main job is recursive dir() to find abundance.h5 files
searches for dir with sampleID, then recursive-searches that for abundance.h5}
\usage{
get_metadata(metadata_csv)
}
\arguments{
\item{metadata_csv}{CSV format file with sample and ontehr columns}
}
\description{
Make metadata from CSV input, including generation of 'path' for abundance.h5 kallisto ouput
main job is recursive dir() to find abundance.h5 files
searches for dir with sampleID, then recursive-searches that for abundance.h5
}
