% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_RNAseq_output.R
\name{get_tx2gen}
\alias{get_tx2gen}
\title{Get annotation in tx2gene format; returns hsapiens_homolog when non-hsapiens input}
\usage{
get_tx2gen(genome_prefix)
}
\arguments{
\item{genome_prefix}{string of a genome in biomart$dataset, suffixed with '_gene_ensembl'}
}
\value{
tx2gene format for sleuth_prep() in get_abundance_tsv()
}
\description{
Get annotation in tx2gene format; returns hsapiens_homolog when non-hsapiens input
}
