% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_module_joins.R
\name{group_agg_multi}
\alias{group_agg_multi}
\title{Aggregate 3, 4 columns based on grouping on 1, 2 columns
NB that table can have multiple annotations in RNAseqR based on contrasts
this function removes the 'per-caller'}
\usage{
group_agg_multi(f, colns = NULL, pattern = NULL)
}
\arguments{
\item{f}{table object with colnames specified in colns}

\item{colns}{colnames of f on which to aggregate}

\item{pattern}{string to grep for colnames on which to operate}
}
\value{
agg_f aggregated gene ID, names table
}
\description{
Aggregate 3, 4 columns based on grouping on 1, 2 columns
NB that table can have multiple annotations in RNAseqR based on contrasts
this function removes the 'per-caller'
}
