% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_module.R
\name{limma_module}
\alias{limma_module}
\title{Limma module taking metadata_csv and raw count data frame to run DE analysis}
\usage{
limma_module(
  count_data,
  anno_tb = NULL,
  tpm_tb = NULL,
  tag = NULL,
  metadata_csv = NULL,
  metadata_design = NULL,
  control_reference = NULL,
  output_dir = NULL,
  delim_samples = "\\\\.",
  run_voom = FALSE
)
}
\arguments{
\item{count_data}{object in matrix format (integers) with rownames as genes and colnames as sample IDs}

\item{anno_tb}{data frame of ensembl_gene_id - external_gene_id mappings for annotation}

\item{tpm_tb}{data frame of tpms to append to results for output}

\item{tag}{string used to prefix output}

\item{metadata_csv}{path to file holding metadata}

\item{metadata_design}{string of design using colnames of metadata_csv BN that final entry is the}

\item{control_reference}{string indicating the 'control' intgroup for last element of design}

\item{output_dir}{path to where output goes}

\item{delim_samples}{character to delimit sample IDs (default: ".")}
}
\value{
nz_fc_co raw count object with no lines summing to zeros
}
\description{
Limma module taking metadata_csv and raw count data frame to run DE analysis
}
