% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne-ap_viper.R
\name{make_aracne_inputs}
\alias{make_aracne_inputs}
\title{Make required inputs for ARACNe-AP which is now supported through RNAseqon}
\usage{
make_aracne_inputs(tpm_tb, metadata, meta_group, tag)
}
\arguments{
\item{tpm_tb}{tpms}

\item{metadata}{metadatas}

\item{meta_group}{metadatas group}

\item{tag}{string}
}
\value{
wide format with higehst exp. when multiple genes same named
}
\description{
Make required inputs for ARACNe-AP which is now supported through RNAseqon
}
