% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_RNAseq_output.R
\name{nf_core_rnaseq_star_salmon_parser}
\alias{nf_core_rnaseq_star_salmon_parser}
\title{Parse Salmon_STAR input as a follow-on from standard nf-core/rnaseq pipeline using STAR-Salmon (v3.3 currently)}
\usage{
nf_core_rnaseq_star_salmon_parser(path = NULL, delim_samples = NULL)
}
\arguments{
\item{path}{string to star_salmon/salmon.merged.gene_counts.tsv output}

\item{delim_samples}{string to delimit sample names (default = "\\.")}
}
\value{
list [\link{c("count_data", "tpm_tb")}]
}
\description{
Parse Salmon_STAR input as a follow-on from standard nf-core/rnaseq pipeline using STAR-Salmon (v3.3 currently)
}
