% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_RNAseq_output.R
\name{obs_raw_widen}
\alias{obs_raw_widen}
\title{Create wide tibble from sleuth object
sums transcripts from a gene-mode sleuth object into counts of those genes}
\usage{
obs_raw_widen(so, which_value, group_col)
}
\arguments{
\item{so}{sleuth object from sleuth_prep in gene_mode}

\item{which_value}{which 'values_from' to use for generating output (default: 'est_counts', else 'tpm')}

\item{group_col}{string of column to use in aggregating ("ensembl_gene_id" if NULL)}
}
\value{
so which has $obs_raw_count$wide and $long
}
\description{
Create wide tibble from sleuth object
sums transcripts from a gene-mode sleuth object into counts of those genes
}
