% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne-ap_viper.R
\name{parse_aracne}
\alias{parse_aracne}
\title{Parse inputs (from aracne-ap_viper Nextflow)}
\usage{
parse_aracne(NETWORK, EXPRMAT, METADATA, TAG, genome_prefix = "hsapiens")
}
\arguments{
\item{NETWORK}{ARACNe_AP network file}

\item{EXPRMAT}{expression matrix in tab-separated format:
header line in format:
geneID sample_1 ... sampleID_n
all other lines in format
geneID_1 value_1 ... value_n etc.}

\item{METADATA}{metadata to discriminate between cohorts for msViper analysis;
tab-separated format; header line format:
sample group
all other lines:
sampleID_1 group_0 etc.}

\item{TAG}{string identifier of the run}

\item{genome_prefix}{string of a genome in biomart$dataset, suffixed with
'_gene_ensembl'}
}
\value{
none, saves two RData files to current dir contianing:
ensembl to external gene mapping (en2ext);
combined.eset, regulonaracne required to run msViper
}
\description{
Parse inputs (from aracne-ap_viper Nextflow)
}
