% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea.R
\name{per_contrast_fgsea_de}
\alias{per_contrast_fgsea_de}
\title{Function to take output table from multiple contrasts
take each level of contrasts and combine those levels into master tables
filter based on min genes per master table}
\usage{
per_contrast_fgsea_de(fgsea_contrast_list, occupancy = 5, output_dir, tag)
}
\arguments{
\item{fgsea_contrast_list}{named list returned from lapply on fgsea_plot()}

\item{occupancy}{\% of genes in the pathway that allow use of the pathway for ssgsea}

\item{output_dir}{path to where output goes}

\item{tag}{string used to prefix output}
}
\value{
table of pathways for each contrast level
}
\description{
Function to take output table from multiple contrasts
take each level of contrasts and combine those levels into master tables
filter based on min genes per master table
}
