% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne-ap_viper.R
\name{run_msviper}
\alias{run_msviper}
\title{Run msViper}
\usage{
run_msviper(TAG, RDATA, genome_prefix, msigdb_species)
}
\arguments{
\item{TAG}{string identifier of the run}

\item{RDATA}{RData file generated by parse_input()}

\item{genome_prefix}{string of a genome in biomart$dataset, suffixed with '_gene_ensembl'}

\item{msigdb_species}{one of msigdbr::msigdbr_species()}
}
\value{
none, saves RData files of output
}
\description{
Run msViper
}
