% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_scripts.R
\name{run_prep_modules_bm}
\alias{run_prep_modules_bm}
\title{Wrap prep and module for data from brucemoran RNAseq_kallisto Nextflow pipeline}
\usage{
run_prep_modules_bm(
  metadata_csv,
  metadata_design,
  tag,
  output_dir,
  control_reference = NULL
)
}
\arguments{
\item{metadata_csv}{path to file holding metadata}

\item{metadata_design}{string of design using colnames of metadata_csv}

\item{tag}{string used to prefix output}

\item{output_dir}{path to where output goes}

\item{control_reference}{string indicating the 'control' intgroup for last element of design}
}
\value{
none rds and tsv files printed
}
\description{
Wrap prep and module for data from brucemoran RNAseq_kallisto Nextflow pipeline
}
