% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_RNAseq_output.R
\name{so_obs_raw_out}
\alias{so_obs_raw_out}
\title{Add wide and long raw count tibbles to sleuth object
sums transcripts from a gene-mode sleuth object into counts of those genes}
\usage{
so_obs_raw_out(so, group_col = NULL, which_value = "est_counts")
}
\arguments{
\item{so}{sleuth object from sleuth_prep in gene_mode}

\item{group_col}{string of column to use in aggregating ("ensembl_gene_id" if NULL)}

\item{which_value}{which 'values_from' to use for generating output (default: 'est_counts', else 'tpm')}
}
\value{
so which has $obs_raw_count$wide and $long
}
\description{
Add wide and long raw count tibbles to sleuth object
sums transcripts from a gene-mode sleuth object into counts of those genes
}
