% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_RNAseq_output.R
\name{so_to_raw_counts}
\alias{so_to_raw_counts}
\title{so to raw_counts_df}
\usage{
so_to_raw_counts(so)
}
\arguments{
\item{so}{from sleuth object treated with so_obs_raw_out}

\item{output}{one of 'count' for counts, 'tpm' for tpm}
}
\description{
so to raw_counts_df
}
