% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgsea.R
\name{ssgsea_pca}
\alias{ssgsea_pca}
\title{Function to run ssGSEA using the GSVA package and return PCA of results}
\usage{
ssgsea_pca(
  pways,
  log2tpm_mat,
  msigdb_cat = "H",
  output_dir,
  contrast,
  metadata,
  plot_out = NULL
)
}
\arguments{
\item{pways}{list of pathway genesets to use in ssGSEA}

\item{log2tpm_mat}{matrix of log2tpm values per sample across genes in gene lists}

\item{msigdb_cat}{one of 'c("H", paste0("C", c(1:7)))', see: gsea-msigdb.org/gsea/msigdb/collections.jsp; if H, loads Process_Category data from Liberzon 2015 paper to colour PCA}

\item{output_dir}{path to where output goes}

\item{contrast}{string used to prefix output naming the contrast being investigated}

\item{metadata}{tibble of sample, \<group\> where group colours the PCA}

\item{plot_out}{logical plot output to screen (for knitr; default NULL)}
}
\value{
ssgsea_ggp_list results table of each geneset and sample
}
\description{
Function to run ssGSEA using the GSVA package and return PCA of results
}
