% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgsea.R
\name{ssgsea_rotationPCA}
\alias{ssgsea_rotationPCA}
\title{Plotting PCA function}
\usage{
ssgsea_rotationPCA(
  ssgsea,
  metadata,
  hallmark_tb = NULL,
  contrast,
  returnData = FALSE
)
}
\arguments{
\item{ssgsea}{output table}

\item{metadata}{tibble of sample, \<group\> where group colours the PCA}

\item{hallmark_tb}{hallmark data with Process_Category column to colour by process}

\item{contrast}{string to title plot}

\item{returnData}{flag to allow data to be returned}
}
\value{
ggp_pca_list ggplot2 object for printing, pca for safe keeping
}
\description{
Plotting PCA function
}
