% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_module.R
\name{BMplotPCA}
\alias{BMplotPCA}
\title{Plotting PCA function}
\usage{
BMplotPCA(x, intgroup = NULL, ntop = 1500, returnData = FALSE)
}
\arguments{
\item{x}{variance stabilized DESeq2 object}

\item{intgroup}{which colname from metadata_csv to be output on plot}

\item{ntop}{top n genes to use by variance}

\item{returnData}{flag to allow data to be returned}
}
\value{
ggp ggplot2 object for printing
}
\description{
Plotting PCA function
}
