% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_RNAseq_output.R
\name{brucemoran_rnaseq_kallisto_parser}
\alias{brucemoran_rnaseq_kallisto_parser}
\title{Take metadata CSV, parse Kallisto output and return a Sleuth object with raw counts}
\usage{
brucemoran_rnaseq_kallisto_parser(
  metadata_csv,
  data_dir = NULL,
  agg_col = "ensembl_gene_id",
  genome_prefix = "hsapiens",
  server_threads = NULL
)
}
\arguments{
\item{metadata_csv}{CSV format file with columns:
'sample' that has exact dir matching which contains 'adundance.h5' (recursively)}

\item{data_dir}{dir from which to recursively search for abundance.h5 for samples}

\item{agg_col}{column on which to aggregate gene_mode}

\item{genome_prefix}{string of a genome in biomart$dataset, suffixed with '_gene_ensembl'}

\item{server_threads}{set to anything but NULL to allow threading}
}
\value{
list object with sleuth object 'so' [\link{1}] and tx2gene annotation used to create it [\link{2}]
}
\description{
Take metadata CSV, parse Kallisto output and return a Sleuth object with raw counts
}
