% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea.R
\name{fgsea_plot}
\alias{fgsea_plot}
\title{Function to run fGSEA using the mSIGDB database from Broad}
\usage{
fgsea_plot(
  res,
  sig_res = NULL,
  msigdb_species = "Homo sapiens",
  msigdb_cat = "H",
  gene_col = NULL,
  padj = 0.01,
  output_dir,
  tag
)
}
\arguments{
\item{res}{output table from 'DESeq2_module()' (all genes not just sig)}

\item{sig_res}{output table from 'DESeq2/limma/edger_module()' or combination thereof (used for table output)}

\item{msigdb_species}{one of msigdbr::msigdbr_species(), default:"Homo sapiens"}

\item{gene_col}{the name of the column in tibble with gene names found in pathways, set to "rownames" if they are the rownames (default)}

\item{padj}{the significance threshold}

\item{output_dir}{path to where output goes}

\item{tag}{string used to prefix output}

\item{rank_col}{colname used to rank data (based on t in limma)}

\item{gene_set}{genes of interest for fgsea (subset to these, use all if not specified)}

\item{msigdb_cats}{one of 'c("H", paste0("C", c(1:8)))', see: gsea-msigdb.org/gsea/msigdb/collections.jsp}
}
\value{
msigdb_fgsea object
}
\description{
Function to run fGSEA using the mSIGDB database from Broad
}
