% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea.R
\name{fgsea_plot}
\alias{fgsea_plot}
\title{Function to run fGSEA using the mSIGDB database from Broad}
\usage{
fgsea_plot(
  res,
  sig_res = NULL,
  msigdb_species = "Homo sapiens",
  msigdb_cat = "H",
  gene_col = NULL,
  rank_col = NULL,
  padj = 0.01,
  output_dir,
  tag,
  contrast,
  plot_out = NULL
)
}
\arguments{
\item{res}{data frame (all genes not just sig)}

\item{sig_res}{data frame of significant results}

\item{msigdb_species}{one of msigdbr::msigdbr_show_species(), default:"Homo sapiens"}

\item{msigdb_cat}{one of 'c("H", paste0("C", c(1:7)))', see: gsea-msigdb.org/gsea/msigdb/collections.jsp}

\item{gene_col}{the name of the column in data frame with gene names found in pathways, set to "rownames" if they are the rownames (default)}

\item{rank_col}{the name of the column in tibble which ranks genes for fgsea (default: "stat" for DESeq2 results; limma - use "t", edgeR - unsure)}

\item{padj}{the significance threshold}

\item{output_dir}{path to where output goes}

\item{tag}{string used to prefix output}

\item{contrast}{string to define the contrast being made, tags output}

\item{plot_out}{logical plot output to screen (for knitr; default NULL)}
}
\value{
msigdb_fgsea object
}
\description{
Function to run fGSEA using the mSIGDB database from Broad
}
