% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_module_joins.R
\name{group_agg_two}
\alias{group_agg_two}
\title{Aggregate 1 columns based on grouping on 2 column}
\usage{
group_agg_two(f, colns = NULL, pattern = NULL)
}
\arguments{
\item{f}{table object with colnames specified in colns}

\item{colns}{colnames of f on which to aggregate}

\item{pattern}{string to grep for colnames on which to operate}
}
\value{
agg_f aggregated gene ID, names table
}
\description{
Aggregate 1 columns based on grouping on 2 column
}
