% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne-ap_viper.R
\name{run_msviper}
\alias{run_msviper}
\title{Run msViper}
\usage{
run_msviper(TAG, RDATA)
}
\arguments{
\item{TAG}{string identifier of the run}

\item{RDATA}{RData file generated by parse_input()}
}
\value{
none, saves two RData files to current dir contianing:
ensembl to external gene mapping (en2ext);
combined.eset, regulonaracne required to run msViper
}
\description{
Run msViper
}
