% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_module_joins.R
\name{venn_3}
\alias{venn_3}
\title{Make venn diagram}
\usage{
venn_3(master_list, contrast, tag, output_dir, padj = 0.01)
}
\arguments{
\item{master_list}{list of 3 DE results}

\item{contrast}{name of contrast to use for table (should be one of 'names(master_list[\link{x}])'')}

\item{tag}{string used to prefix output}

\item{output_dir}{path to where output goes}

\item{padj}{significance threshold below which to determine significance (adjusted by FDR)}
}
\value{
venn diagram of 3 DE callers
}
\description{
Make venn diagram
}
